requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
numpy
transformers
torch
